+++ src/app.ts
@@ -9,9 +9,9 @@ import { DEV_CONFIG, PROD_CONFIG } from '#/constants/setup';
 import { NotFoundError } from '#/errors/definedErrors';
 import errorHandler from '#/errors/errorHandler';
 import router from '#/routes';
-import { logger } from '#/libs/logger/winstonLogger';
+import { logger } from '#/libs/logger';
 import mongoConnection from '#/utils/connectMongoDB';
-import { stream } from '#/libs/logger/winstonLogger';
+import { stream } from '#/libs/logger';
 
 dotenv.config();
