+++ src/libs/aws-sdk/s3.ts
@@ -6,7 +6,7 @@ import {
 import fs from 'fs';
 
 import { InternalServerError } from '#/errors/definedErrors';
-import { logger } from '#/libs/logger/winstonLogger';
+import { logger } from '#/libs/logger';
 
 const storage = new S3Client({
   region: 'ap-northeast-2',
